# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Megacoop API Documentation",
        "description": "Cooperative",
        "contact": {
            "email": "development@leverpay.io"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Test server"
        }
    ],
    "paths": {
...
        "/api/v1/user/complete-signup": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Complete user signup after code validation",
                "description": "Complete Signup",
                "operationId": "09f4bf56957b6f0de511cdd92abd8f32",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "gender",
                                    "address",
                                    "dob",
                                    "code",
                                    "email"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "4Trdsgq7"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "08012345678"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ],
                                        "example": "male"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Lagos Street, Nigeria"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-06-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Signup completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Signup completed successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "middle_name": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Smith"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "08012345678"
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "example": "male"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "123 Lagos Street, Nigeria"
                                                },
                                                "dob": {
                                                    "type": "string",
                                                    "example": "1995-06-15"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8000/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    ]
}